Quick Sort
Implement a Quick Sort algorithm to sort a given set of elements and determine the time required to sort the elements. Repeat the experiment for different values of n, the number of
elements in the list to be sorted. The elements can be read from the user or can be generated using the random number generator